ACTIN-1371: More renames of hotspot to variant
