+++ algo/src/main/java/com/hartwig/serve/extraction/EventExtractorFactory.java
@@ -16,11 +16,11 @@ import com.hartwig.serve.extraction.copynumber.CopyNumberExtractor;
 import com.hartwig.serve.extraction.exon.ExonExtractor;
 import com.hartwig.serve.extraction.fusion.FusionExtractor;
 import com.hartwig.serve.extraction.gene.GeneLevelExtractor;
-import com.hartwig.serve.extraction.hotspot.VariantExtractor;
 import com.hartwig.serve.extraction.immuno.ImmunoHLAExtractor;
 import com.hartwig.serve.extraction.util.DriverInconsistencyMode;
 import com.hartwig.serve.extraction.util.GeneChecker;
 import com.hartwig.serve.extraction.util.MutationTypeFilterAlgo;
+import com.hartwig.serve.extraction.variant.VariantExtractor;
 import com.hartwig.serve.refgenome.RefGenomeResource;
 
 import org.jetbrains.annotations.NotNull;
