+++ algo/src/main/java/com/hartwig/serve/extraction/ExtractionFunctions.java
@@ -23,7 +23,7 @@ import com.hartwig.serve.extraction.events.EventInterpretation;
 import com.hartwig.serve.extraction.exon.ExonConsolidation;
 import com.hartwig.serve.extraction.fusion.FusionConsolidation;
 import com.hartwig.serve.extraction.gene.GeneConsolidation;
-import com.hartwig.serve.extraction.hotspot.KnownHotspotConsolidation;
+import com.hartwig.serve.extraction.variant.KnownHotspotConsolidation;
 
 import org.apache.commons.compress.utils.Lists;
 import org.apache.logging.log4j.LogManager;
