+++ algo/src/main/java/com/hartwig/serve/extraction/variant/HumanChromosome.java
@@ -1,4 +1,4 @@
-package com.hartwig.serve.extraction.hotspot;
+package com.hartwig.serve.extraction.variant;
 
 import com.hartwig.serve.common.RefGenomeFunctions;
