+++ algo/src/main/java/com/hartwig/serve/extraction/variant/KnownHotspotVCF.java
@@ -1,4 +1,4 @@
-package com.hartwig.serve.extraction.hotspot;
+package com.hartwig.serve.extraction.variant;
 
 import java.io.File;
 import java.io.IOException;
