+++ algo/src/main/java/com/hartwig/serve/extraction/variant/MitochondrialChromosome.java
@@ -1,4 +1,4 @@
-package com.hartwig.serve.extraction.hotspot;
+package com.hartwig.serve.extraction.variant;
 
 import org.jetbrains.annotations.NotNull;
