+++ algo/src/main/java/com/hartwig/serve/extraction/variant/ProteinResolver.java
@@ -1,4 +1,4 @@
-package com.hartwig.serve.extraction.hotspot;
+package com.hartwig.serve.extraction.variant;
 
 import java.util.List;
 import java.util.Set;
