+++ algo/src/main/java/com/hartwig/serve/extraction/variant/tools/AnnotatedHotspotVCFPrinter.java
@@ -1,4 +1,4 @@
-package com.hartwig.serve.extraction.hotspot.tools;
+package com.hartwig.serve.extraction.variant.tools;
 
 import static htsjdk.tribble.AbstractFeatureReader.getFeatureReader;
