+++ algo/src/main/java/com/hartwig/serve/sources/docm/DocmExtractor.java
@@ -13,9 +13,9 @@ import com.hartwig.serve.datamodel.molecular.hotspot.ImmutableKnownHotspot;
 import com.hartwig.serve.datamodel.molecular.hotspot.KnownHotspot;
 import com.hartwig.serve.extraction.ExtractionResult;
 import com.hartwig.serve.extraction.ImmutableExtractionResult;
-import com.hartwig.serve.extraction.hotspot.KnownHotspotConsolidation;
-import com.hartwig.serve.extraction.hotspot.ProteinResolver;
-import com.hartwig.serve.extraction.hotspot.Variant;
+import com.hartwig.serve.extraction.variant.KnownHotspotConsolidation;
+import com.hartwig.serve.extraction.variant.ProteinResolver;
+import com.hartwig.serve.extraction.variant.Variant;
 import com.hartwig.serve.util.ProgressTracker;
 
 import org.jetbrains.annotations.NotNull;
