+++ algo/src/main/java/com/hartwig/serve/transvar/TransvarInterpreter.java
@@ -12,8 +12,8 @@ import com.google.common.collect.Lists;
 import com.hartwig.serve.common.RefGenomeFunctions;
 import com.hartwig.serve.common.ensemblcache.Strand;
 import com.hartwig.serve.datamodel.RefGenome;
-import com.hartwig.serve.extraction.hotspot.ImmutableVariant;
-import com.hartwig.serve.extraction.hotspot.Variant;
+import com.hartwig.serve.extraction.variant.ImmutableVariant;
+import com.hartwig.serve.extraction.variant.Variant;
 import com.hartwig.serve.transvar.datamodel.TransvarAnnotation;
 import com.hartwig.serve.transvar.datamodel.TransvarComplexInsertDelete;
 import com.hartwig.serve.transvar.datamodel.TransvarDeletion;
