+++ algo/src/test/java/com/hartwig/serve/extraction/variant/KnownHotspotVCFTest.java
@@ -1,4 +1,4 @@
-package com.hartwig.serve.extraction.hotspot;
+package com.hartwig.serve.extraction.variant;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
